# backend/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OLB Transfers API",
    "description" : "Modernized Bank of America Online Banking Transfers API. Session-cookie auth: POST /api/login, then call /api/secure/** with the SESSION cookie and X-XSRF-TOKEN header (value of the XSRF-TOKEN cookie) on mutating requests.",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "http://localhost:44917",
    "description" : "Generated server url"
  } ],
...
  "paths" : {
...
    "/api/login" : {
      "post" : {
        "tags" : [ "Session" ],
        "summary" : "Sign in",
        "description" : "Legacy /login.do. Case-insensitive user id, bcrypt password, 3-strike permanent lockout (AC-01..AC-05). A new session is created on success (fixation protection).",
        "operationId" : "login",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LoginRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Signed in; session cookie set",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "LOGIN_REQUIRED / LOGIN_INVALID / LOGIN_LOCKED with the legacy message",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/csrf" : {
      "get" : {
        "tags" : [ "Session" ],
        "summary" : "CSRF token",
        "description" : "Returns the XSRF-TOKEN value (also set as a readable cookie on every response). Send it back as X-XSRF-TOKEN on POST/PUT/DELETE to /api/secure/**.",
        "operationId" : "csrf",
        "parameters" : [ {
          "name" : "token",
          "in" : "query",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/CsrfToken"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "string"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "ApiError" : {
        "type" : "object",
        "description" : "Error envelope. `message` is the verbatim legacy user-facing text; `code` is stable for clients.",
        "properties" : {
          "ok" : {
            "type" : "boolean",
            "example" : false
          },
          "status" : {
            "type" : "integer",
            "format" : "int32",
            "example" : 422
          },
          "code" : {
            "type" : "string",
            "example" : "XFR_PERTXN"
          },
          "message" : {
            "type" : "string",
            "example" : "This transfer exceeds the per-transfer limit of $3,500.00 for your relationship tier."
          }
        }
      },
...
      "LoginRequest" : {
        "type" : "object",
        "properties" : {
          "userId" : {
            "type" : "string",
            "example" : "demo.user"
          },
          "password" : {
            "type" : "string",
            "example" : "Password1"
          },
          "saveUserId" : {
            "type" : "boolean",
            "description" : "Legacy 'Save user ID' checkbox: sets/clears the olb_uid cookie (AC-06)",
            "example" : true
          }
        }
      },
      "CodeLabel" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "label" : {
            "type" : "string"
          }
        }
      },
      "MeResponse" : {
        "type" : "object",
        "properties" : {
          "customerId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "userId" : {
            "type" : "string"
          },
          "firstName" : {
            "type" : "string"
          },
          "lastName" : {
            "type" : "string"
          },
          "displayName" : {
            "type" : "string"
          },
          "tierCode" : {
            "type" : "string"
          },
          "tierLabel" : {
            "type" : "string"
          },
          "tierOptions" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CodeLabel"
            }
          },
          "frequencyOptions" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CodeLabel"
            }
          },
          "deliveryOptions" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CodeLabel"
            }
          },
          "lastLogin" : {
            "type" : "string",
            "format" : "date-time"
          },
          "sessionTimeoutSeconds" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "CsrfToken" : {
        "type" : "object",
        "properties" : {
          "parameterName" : {
            "type" : "string"
          },
          "token" : {
            "type" : "string"
          },
          "headerName" : {
            "type" : "string"
          }
        }
      }
    }
  }
}